pokemonapp/requirements.txt
Django~=3.1.2
djangorestframework~=3.12.4
requests~=2.23.0